ref_owlvit/requirements.txt
torch
torchvision
transformers
opencv-python-headless
Pillow
numpy<2.0,>=1.21.0
tqdm
PyMuPDF
pycocotools